ref(embeddings): refactor logic
