+++ src/sentry/api/endpoints/project_backfill_similar_issues_embeddings_records.py
@@ -8,7 +8,9 @@ from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import region_silo_endpoint
 from sentry.api.bases.project import ProjectEndpoint
 from sentry.auth.superuser import is_active_superuser
-from sentry.tasks.backfill_seer_grouping_records import backfill_seer_grouping_records
+from sentry.tasks.embeddings_grouping.backfill_seer_grouping_records_for_project import (
+    backfill_seer_grouping_records_for_project,
+)
 
 
 @region_silo_endpoint
@@ -41,5 +43,7 @@ class ProjectBackfillSimilarIssuesEmbeddingsRecords(ProjectEndpoint):
         if request.data.get("only_delete"):
             only_delete = True
 
-        backfill_seer_grouping_records.delay(project.id, last_processed_index, dry_run, only_delete)
+        backfill_seer_grouping_records_for_project.delay(
+            project.id, last_processed_index, dry_run, only_delete
+        )
         return Response(status=204)
